Move Dictionary into types.ts
