--- src/lib/customTypes.ts
@@ -1,5 +1,0 @@
-/* eslint-disable no-unused-vars */
-/* eslint-disable @typescript-eslint/no-namespace */
-export type Dictionary<T> = {
-    [key: string]: T;
-};
